services/docreader-http/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
minio>=7.2.0
pydantic-settings>=2.6.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-pptx>=1.0.0
pillow>=10.0.0
# 与 WeKnora docreader 对齐：优先 MarkItDown 转换 Office/PDF 等，再回退内置解析器
markitdown[docx,pdf,pptx,xlsx]>=0.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
